requirements.txt
python-docx
pypdf
openpyxl
python-pptx
